Backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-bcrypt==1.0.1
PyJWT==2.10.1
mysql-connector-python==9.2.0
python-dotenv==1.0.1
pytesseract==0.3.13
Pillow==11.1.0
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
twilio
